tuned PIDs (spindexer not fully tuned)
